reduce service worker bundle size
