+++ packages/recoil/src/types.ts
@@ -1,15 +1,4 @@
-import {
-  type Blockchain,
-  makeUrl,
-  TAB_APPS,
-  TAB_BALANCES,
-  TAB_MESSAGES,
-  TAB_NFTS,
-  TAB_NOTIFICATIONS,
-  TAB_RECENT_ACTIVITY,
-  TAB_SWAP,
-  TAB_TOKENS,
-} from "@coral-xyz/common";
+import { type Blockchain } from "@coral-xyz/common";
 import type { BigNumber } from "ethers";
 import type { RecoilValueReadOnly } from "recoil";
 
@@ -84,32 +73,6 @@ export interface TokenMetadata {
   name: string;
   image: string;
   symbol: string;
-}
-
-export const TABS = [
-  [TAB_BALANCES, "Balances"],
-  [TAB_NFTS, "Nfts"],
-  [TAB_SWAP, "Swap"],
-  [TAB_APPS, "Apps"],
-  [TAB_MESSAGES, "Messages"],
-  [TAB_RECENT_ACTIVITY, "Recent Activity"],
-  [TAB_NOTIFICATIONS, "Notifications"],
-  [TAB_TOKENS, "Tokens"],
-];
-
-export function makeDefaultNav() {
-  const defaultNav: any = {
-    activeTab: TAB_TOKENS,
-    data: {},
-  };
-  TABS.forEach(([tabName, tabTitle]) => {
-    defaultNav.data[tabName] = {
-      id: tabName,
-      urls: [makeUrl(tabName, { title: tabTitle, props: {} })],
-      ref: tabName === "balances" ? "tokens" : undefined,
-    };
-  });
-  return defaultNav;
 }
 
 export type ExtractRecoilType<P> = P extends RecoilValueReadOnly<infer T>
